github-upload/backend/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Processamento de dados
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2

# Análise de texto e NLP
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# OCR
pytesseract==0.3.10
easyocr==1.7.0
Pillow==10.1.0

# Processamento de documentos
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.0

# Banco de dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
neo4j==5.14.1

# Análise e clustering
scikit-learn==1.3.2
networkx==3.2.1

# Utilitários
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.0

# IA e LLM (opcional)
openai==1.3.7
langchain==0.0.340 